sentrycore: wrap error with log message (#2)
Right now, error messages are super non-descript because most context is in the log message, not in the error message. I think it's useful to have the log message here because it also matters where each error is being reported from, so grouping by log messages is something we probably want.

And of course we add the scope, because we want those to become important keys in navigating observability reports.