gomod: lower minimum version to go1.17
We are currently investigating a memory regression in zoekt which seems
to be linked to go1.18. As such we have been experimenting with go1.17
to compare. But we can't use this module on go1.17. However, there is
nothing in this module that requires go1.18, so we can lower the minimum
version. The only code change was to stop using the type alias "any".

Test Plan: go test