introduce SRC_LOG_SCOPE_LEVEL (#46)
SRC_LOG_SCOPE_LEVEL allows operators to override the log level for
specific loggers and there children. It is an comma seperated
environment variable which takes a logger name and level. If a log
message matches the logger name (or a parent of it does) AND the log
level allows it then it will always be logged. A parent is defined by
how we join scoped loggers with a ".".

The motivation for this is in production we log with very low verbosity,
but when testing a feature it can be useful to turn up the verbosity for
specific logs. For example while I was developing hybrid searcher it
would of been great to make just that debug, but avoid debug for
everything else due to the noise.

Test Plan: Added unit tests which test from the envar level.