outputcore: correctly use dev mode to apply configuration (#50)
`encoders.BuildEncoder` relies on the `development bool` argument to assess whether to apply our dev config for easier-to-read output. It looks like we no longer call `BuildEncoder` with anything except `development: false` - I don't recall what it used to be, but setting it to `globallogger.DevMode()` should fix it. We can't place this in `BuildEncoder` itself because that would cause an import cycle between package `encoders` and `globallogger`.

Since we have `(*outputSink).development` available, however, I've opted to prop-drill it in to `NewCore` from the top-level variable. `development` is a bit problematic to work with - we might want to better refactor the interactions here at some point.

Test plan: I added a sanity-check program, `cmd/example`, to do some quick-and-easy visual confirmation of log output.