feat(overrides): updated overrideCore to allow filtering out logs at a higher level (#71)
As per our discussion in
[slack](https://sourcegraph.slack.com/archives/C04MYFW01NV/p1741728615123969),
I think it would be nice to allow override logs to raise the level of
severity. To that end, my proposed changes are:
* Always return an override core if overrides are provided
* Check overrides first in override core, and then fall back to base
levels

I acknowledge this does introduce a performance hit when
SRC_LOG_SCOPE_LEVEL is defined, but I doubt it's significant. If we're
concerned, I could put this behavior behind an env var like
SRC_LOG_OVERRIDES_CAN_INCREASE_SEVERITY (not the punchiest, but gets the
point across). However, I'd personally prefer if it was the default
because I was surprised by the current behavior and clearly thought it
was a bug.

## Test Plan
Updated and added unit test

Closes
https://linear.app/sourcegraph/issue/DINF-819/sg-start-not-respecting-src-log-scope-level-variable